Refactor browser extension to use shared vault models (#900)
