+++ apps/browser-extension/src/entrypoints/contentScript/Popup.ts
@@ -6,9 +6,9 @@ import { fillCredential } from '@/entrypoints/contentScript/Form';
 import { DISABLED_SITES_KEY, TEMPORARY_DISABLED_SITES_KEY, GLOBAL_AUTOFILL_POPUP_ENABLED_KEY, VAULT_LOCKED_DISMISS_UNTIL_KEY, LAST_CUSTOM_EMAIL_KEY, LAST_CUSTOM_USERNAME_KEY } from '@/utils/Constants';
 import { FormDetector } from '@/utils/formDetector/FormDetector';
 import { CreateIdentityGenerator } from '@/utils/shared/identity-generator';
+import type { Credential } from '@/utils/shared/models/vault';
 import { CreatePasswordGenerator, PasswordGenerator } from '@/utils/shared/password-generator';
 import { SqliteClient } from '@/utils/SqliteClient';
-import { Credential } from '@/utils/types/Credential';
 import { CredentialsResponse } from '@/utils/types/messaging/CredentialsResponse';
 import { PasswordSettingsResponse } from '@/utils/types/messaging/PasswordSettingsResponse';
 import { StringResponse } from '@/utils/types/messaging/StringResponse';
