+++ apps/browser-extension/src/entrypoints/popup/components/CredentialDetails/AliasBlock.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { FormInputCopyToClipboard } from '@/entrypoints/popup/components/FormInputCopyToClipboard';
 
 import { IdentityHelperUtils } from '@/utils/shared/identity-generator';
-import { Credential } from '@/utils/types/Credential';
+import type { Credential } from '@/utils/shared/models/vault';
 
 type AliasBlockProps = {
   credential: Credential;
