+++ apps/browser-extension/src/entrypoints/popup/pages/CredentialsList.tsx
@@ -8,8 +8,8 @@ import { useDb } from '@/entrypoints/popup/context/DbContext';
 import { useLoading } from '@/entrypoints/popup/context/LoadingContext';
 import { useWebApi } from '@/entrypoints/popup/context/WebApiContext';
 
+import type { Credential } from '@/utils/shared/models/vault';
 import type { VaultResponse } from '@/utils/shared/models/webapi';
-import { Credential } from '@/utils/types/Credential';
 
 import { useMinDurationLoading } from '@/hooks/useMinDurationLoading';
