--- apps/browser-extension/src/utils/types/EncryptionKey.ts
@@ -1,9 +1,0 @@
-/**
- * Encryption key SQLite database type.
- */
-export type EncryptionKey = {
-    Id: string;
-    PublicKey: string;
-    PrivateKey: string;
-    IsPrimary: boolean;
-}
