+++ apps/browser-extension/src/utils/types/errors/ApiAuthError.ts
@@ -4,7 +4,7 @@
 export class ApiAuthError extends Error {
   /**
    * Creates a new instance of ApiAuthError.
-   * 
+   *
    * @param message - The error message.
    */
   public constructor(message: string) {
